refactor(providers): mirror usage query core across both frontends
将 usageQuery 的状态归约/协调器/hook 逻辑抽入 usageQueryCore.ts,
按镜像纪律加入 mirror-manifest 逐字节强制;两端 usageQuery.ts 只保留
平台传输适配(GUI 走 Tauri invoke,WebUI 走 Gateway WebSocket 桥接)。
providerUtils.ts 两端本已逐字节一致,一并入 manifest 防漂移。
