+++ scripts/mirror-manifest.json
@@ -39,6 +39,7 @@
     "lib/normalizeLatexDelimiters.ts",
     "lib/providers/customHeaders.ts",
     "lib/providers/proxy.ts",
+    "lib/providers/usageQueryCore.ts",
     "lib/hubFetch.ts",
     "lib/runtimeEnv.ts",
     "components/workspace-editor/workspaceMarkdownAssets.ts",
@@ -106,6 +107,7 @@
     "lib/shared/fontFamily.ts",
     "pages/settings/SystemToolsSection.tsx",
     "pages/settings/SshSection.tsx",
+    "pages/settings/providerUtils.ts",
     "lib/terminal/types.ts",
     "lib/terminal/sshLocalForwardTypes.ts",
     "components/project-tools/SshPortForwardDialog.tsx",
